solving a bug with delta V

+++ src/main/java/com/rae/creatingspace/content/rocket/contraption/behaviour/interaction/FlightRecorderInteraction.java
@@ -44,7 +44,7 @@ public class FlightRecorderInteraction extends MovingInteractionBehaviour {
 
                             );
                             for (TagKey<Fluid> fluidTagKey : lastAssemblyData.propellantStatusData().consumedMassForEachPropellant().keySet()) {
-                                Integer consumedMass = lastAssemblyData.propellantStatusData().consumedMassForEachPropellant().get(fluidTagKey);
+                                Float consumedMass = lastAssemblyData.propellantStatusData().consumedMassForEachPropellant().get(fluidTagKey);
                                 Integer fluidMass = lastAssemblyData.propellantStatusData().massForEachPropellant().get(fluidTagKey);
 
                                 if (fluidMass == null) {
