+++ src/main/java/com/rae/creatingspace/content/rocket/flight_recorder/FlightRecorderBlockEntity.java
@@ -90,7 +90,7 @@ public class FlightRecorderBlockEntity extends KineticBlockEntity implements IHa
                             )
                             .forGoggles(tooltip, 1);
                     for (TagKey<Fluid> fluidTagKey : lastAssemblyData.propellantStatusData().consumedMassForEachPropellant().keySet()) {
-                        Integer consumedMass = lastAssemblyData.propellantStatusData().consumedMassForEachPropellant().get(fluidTagKey);
+                        Float consumedMass = lastAssemblyData.propellantStatusData().consumedMassForEachPropellant().get(fluidTagKey);
                         Integer fluidMass = lastAssemblyData.propellantStatusData().massForEachPropellant().get(fluidTagKey);
 
                         if (fluidMass == null) {
